+++ controllers/config/config.go
@@ -34,16 +34,19 @@ const (
 )
 
 var Stage = StageProduction
+var log = ctrl.Log.WithName("config")
 
-var (
+const (
+	envVarProxyUrl                                = "PROXY_URL"
 	envVarNamespace                               = "NAMESPACE"
 	envVarDoguRegistryEndpoint                    = "DOGU_REGISTRY_ENDPOINT"
 	envVarDoguRegistryUsername                    = "DOGU_REGISTRY_USERNAME"
 	envVarDoguRegistryPassword                    = "DOGU_REGISTRY_PASSWORD"
 	envVarDoguRegistryURLSchema                   = "DOGU_REGISTRY_URLSCHEMA"
 	envVarNetworkPolicyEnabled                    = "NETWORK_POLICIES_ENABLED"
+	envVarAuthRegistrationEnabled                 = "AUTH_REGISTRATION_ENABLED"
+	envVarDisablePostfixDependencyCheck           = "DISABLE_POSTFIX_DEPENDENCY_CHECK"
 	envVarRequeueTimeForDoguResourceInNanoseconds = "REQUEUE_TIME_FOR_DOGU_RESOURCE_IN_NANOSECONDS"
-	log                                           = ctrl.Log.WithName("config")
 )
 
 // DoguRegistryData contains all necessary data for the dogu registry.
@@ -64,6 +67,12 @@ type OperatorConfig struct {
 	Version *core.Version `json:"version"`
 	// NetworkPoliciesEnabled defines whether network policies should be created for dogus and their dependencies
 	NetworkPoliciesEnabled bool `json:"network_policies_enabled"`
+	// AuthRegistrationEnabled defines whether the operator should manage AuthRegistration CRs for v2 dogus.
+	AuthRegistrationEnabled bool `json:"auth_registration_enabled"`
+	// DisablePostfixDependencyCheck defines whether the operator should validate dependencies on postfix.
+	// If set to false, the operator will assume that postfix is installed as a normal dogu and will validate the dependencies accordingly.
+	// If set to true, the operator will assume that postfix is installed as a component and will not validate the dependencies.
+	DisablePostfixDependencyCheck bool `json:"disable_postfix_dependency_check"`
 	// RequeueTimeForDoguReconciler defines the requeue time for the dogu reconciler
 	RequeueTimeForDoguReconciler time.Duration `json:"requeue_time_for_dogu_reconciler"`
 }
@@ -72,7 +81,7 @@ type Version string
 
 // NewOperatorConfig creates a new operator config by reading values from the environment variables
 func NewOperatorConfig(version Version) (*OperatorConfig, error) {
-	stage, err := getEnvVar(StageEnvironmentVariable)
+	stage, err := getRequiredEnvVar(StageEnvironmentVariable)
 	if err != nil {
 		log.Error(err, "Error reading stage environment variable. Use Stage production")
 	}
@@ -107,16 +116,18 @@ func NewOperatorConfig(version Version) (*OperatorConfig, error) {
 	log.Info(fmt.Sprintf("Found stored dogu reconciler requeue time! Using requeue time %s", doguReconcilerRequeueTime.String()))
 
 	return &OperatorConfig{
-		Namespace:                    namespace,
-		DoguRegistry:                 doguRegistryData,
-		Version:                      &parsedVersion,
-		NetworkPoliciesEnabled:       getNetworkPoliciesEnabled(),
-		RequeueTimeForDoguReconciler: doguReconcilerRequeueTime,
+		Namespace:                     namespace,
+		DoguRegistry:                  doguRegistryData,
+		Version:                       &parsedVersion,
+		NetworkPoliciesEnabled:        getNetworkPoliciesEnabled(),
+		AuthRegistrationEnabled:       getAuthRegistrationEnabled(),
+		DisablePostfixDependencyCheck: getDisablePostfixDependencyCheck(),
+		RequeueTimeForDoguReconciler:  doguReconcilerRequeueTime,
 	}, nil
 }
 
 func readNamespace() (string, error) {
-	namespace, err := getEnvVar(envVarNamespace)
+	namespace, err := getRequiredEnvVar(envVarNamespace)
 	if err != nil {
 		return "", newEnvVarError(envVarNamespace, err)
 	}
@@ -125,7 +136,7 @@ func readNamespace() (string, error) {
 }
 
 func readDoguReconcilerRequeueTime() (time.Duration, error) {
-	requeueTimeString, err := getEnvVar(envVarRequeueTimeForDoguResourceInNanoseconds)
+	requeueTimeString, err := getRequiredEnvVar(envVarRequeueTimeForDoguResourceInNanoseconds)
 	if err != nil {
 		return defaultRequeueTime, newEnvVarError(envVarNamespace, err)
 	}
@@ -137,24 +148,24 @@ func readDoguReconcilerRequeueTime() (time.Duration, error) {
 }
 
 func readDoguRegistryData() (DoguRegistryData, error) {
-	endpoint, err := getEnvVar(envVarDoguRegistryEndpoint)
+	endpoint, err := getRequiredEnvVar(envVarDoguRegistryEndpoint)
 	if err != nil {
 		return DoguRegistryData{}, newEnvVarError(envVarDoguRegistryEndpoint, err)
 	}
 	// remove tailing slash
 	endpoint = strings.TrimSuffix(endpoint, "/")
 
-	username, err := getEnvVar(envVarDoguRegistryUsername)
+	username, err := getRequiredEnvVar(envVarDoguRegistryUsername)
 	if err != nil {
 		return DoguRegistryData{}, newEnvVarError(envVarDoguRegistryUsername, err)
 	}
 
-	password, err := getEnvVar(envVarDoguRegistryPassword)
+	password, err := getRequiredEnvVar(envVarDoguRegistryPassword)
 	if err != nil {
 		return DoguRegistryData{}, newEnvVarError(envVarDoguRegistryPassword, err)
 	}
 
-	urlschema, err := getEnvVar(envVarDoguRegistryURLSchema)
+	urlschema, err := getRequiredEnvVar(envVarDoguRegistryURLSchema)
 	if err != nil {
 		log.Info(envVarDoguRegistryURLSchema + " not set, using default")
 		urlschema = "default"
@@ -168,7 +179,7 @@ func readDoguRegistryData() (DoguRegistryData, error) {
 	}, nil
 }
 
-func getEnvVar(name string) (string, error) {
+func getRequiredEnvVar(name string) (string, error) {
 	ns, found := os.LookupEnv(name)
 	if !found {
 		return "", fmt.Errorf("environment variable %s must be set", name)
@@ -191,7 +202,7 @@ func (o *OperatorConfig) GetRemoteConfiguration() (*core.Remote, error) {
 		endpoint = strings.TrimSuffix(endpoint, "dogus")
 	}
 
-	proxyURL, found := os.LookupEnv("PROXY_URL")
+	proxyURL, found := os.LookupEnv(envVarProxyUrl)
 	proxySettings := core.ProxySettings{}
 	if found && len(proxyURL) > 0 {
 		var err error
@@ -247,10 +258,10 @@ func newEnvVarError(envVar string, err error) error {
 }
 
 func getNetworkPoliciesEnabled() bool {
-	netPolEnabledStr, err := getEnvVar(envVarNetworkPolicyEnabled)
-	if err != nil {
-		log.Error(fmt.Errorf("failed to read %s from environment: %w", envVarNetworkPolicyEnabled, err), "Enabling network policies by default")
-		return true
+	netPolEnabledStr, found := os.LookupEnv(envVarNetworkPolicyEnabled)
+	if !found {
+		log.Info(fmt.Sprintf("Environment variable %s not set. Enabling network policies by default", envVarNetworkPolicyEnabled))
+		return false
 	}
 
 	netPolEnabled, err := strconv.ParseBool(netPolEnabledStr)
@@ -262,8 +273,40 @@ func getNetworkPoliciesEnabled() bool {
 	return netPolEnabled
 }
 
+func getAuthRegistrationEnabled() bool {
+	authRegistrationEnabledStr, found := os.LookupEnv(envVarAuthRegistrationEnabled)
+	if !found {
+		log.Info(fmt.Sprintf("Environment variable %s not set. Disabling auth registration by default", envVarAuthRegistrationEnabled))
+		return false
+	}
+
+	authRegistrationEnabled, err := strconv.ParseBool(authRegistrationEnabledStr)
+	if err != nil {
+		log.Error(fmt.Errorf("failed to parse value of environment variable %s: %w", envVarAuthRegistrationEnabled, err), "Disabling auth registration by default")
+		return false
+	}
+
+	return authRegistrationEnabled
+}
+
+func getDisablePostfixDependencyCheck() bool {
+	disablePostfixDependencyCheckStr, found := os.LookupEnv(envVarDisablePostfixDependencyCheck)
+	if !found {
+		log.Info(fmt.Sprintf("Environment variable %s not set. Leaving postfix dependency check enabled", envVarDisablePostfixDependencyCheck))
+		return false
+	}
+
+	disablePostfixDependencyCheck, err := strconv.ParseBool(disablePostfixDependencyCheckStr)
+	if err != nil {
+		log.Error(fmt.Errorf("failed to parse value of environment variable %s: %w", envVarDisablePostfixDependencyCheck, err), "Leaving postfix dependency check enabled")
+		return false
+	}
+
+	return disablePostfixDependencyCheck
+}
+
 func GetStage() (string, error) {
-	stage, err := getEnvVar(StageEnvironmentVariable)
+	stage, err := getRequiredEnvVar(StageEnvironmentVariable)
 	if err != nil {
 		return "", err
 	}