+++ controllers/health/doguChecker.go
@@ -11,6 +11,8 @@ import (
 	k8sv2 "github.com/cloudogu/k8s-dogu-lib/v2/api/v2"
 	doguClient "github.com/cloudogu/k8s-dogu-lib/v2/client"
 	"github.com/cloudogu/k8s-dogu-operator/v3/controllers/cesregistry"
+	"github.com/cloudogu/k8s-dogu-operator/v3/controllers/config"
+	doguDepencency "github.com/cloudogu/k8s-dogu-operator/v3/controllers/dependency"
 
 	metav1api "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
@@ -37,16 +39,20 @@ func (dhe *DoguHealthError) Error() string {
 }
 
 // NewDoguChecker creates a checker for dogu health.
-func NewDoguChecker(ecosystemClient doguClient.EcoSystemV2Interface, localFetcher cesregistry.LocalDoguFetcher) DoguHealthChecker {
+func NewDoguChecker(operatorConfig *config.OperatorConfig, ecosystemClient doguClient.EcoSystemV2Interface, localFetcher cesregistry.LocalDoguFetcher) DoguHealthChecker {
 	return &doguChecker{
-		ecosystemClient:   ecosystemClient,
-		doguLocalRegistry: localFetcher,
+		ecosystemClient:               ecosystemClient,
+		doguLocalRegistry:             localFetcher,
+		authRegistrationEnabled:       operatorConfig.AuthRegistrationEnabled,
+		disablePostfixDependencyCheck: operatorConfig.DisablePostfixDependencyCheck,
 	}
 }
 
 type doguChecker struct {
-	ecosystemClient   doguClient.EcoSystemV2Interface
-	doguLocalRegistry localDoguFetcher
+	ecosystemClient               doguClient.EcoSystemV2Interface
+	doguLocalRegistry             localDoguFetcher
+	authRegistrationEnabled       bool
+	disablePostfixDependencyCheck bool
 }
 
 // CheckByName returns nil if the dogu resource's health status says it's available.
@@ -92,9 +98,18 @@ func (dc *doguChecker) checkMandatoryRecursive(ctx context.Context, localDogu *c
 	var errs []error
 
 	for _, dependency := range localDogu.GetDependenciesOfType(core.DependencyTypeDogu) {
-		if dependency.Name == "nginx" || dependency.Name == "registrator" {
+		if dependency.Name == doguDepencency.LegacyDoguNginx || dependency.Name == doguDepencency.LegacyDoguRegistrator {
 			continue
 		}
+
+		if dc.authRegistrationEnabled && dependency.Name == doguDepencency.ComponentDoguCas {
+			continue
+		}
+
+		if dc.disablePostfixDependencyCheck && dependency.Name == doguDepencency.ComponentDoguPostfix {
+			continue
+		}
+
 		localDependencyDoguName := types.NamespacedName{Name: dependency.Name, Namespace: namespace}
 
 		dependencyDogu, err := dc.doguLocalRegistry.FetchInstalled(ctx, cescommons.SimpleName(dependency.Name))
@@ -124,6 +139,14 @@ func (dc *doguChecker) checkOptionalRecursive(ctx context.Context, localDogu *co
 
 	for _, dependency := range localDogu.GetOptionalDependenciesOfType(core.DependencyTypeDogu) {
 		if dependency.Name == "nginx" || dependency.Name == "registrator" {
+			continue
+		}
+
+		if dc.authRegistrationEnabled && dependency.Name == "cas" {
+			continue
+		}
+
+		if dc.disablePostfixDependencyCheck && dependency.Name == "postfix" {
 			continue
 		}