+++ controllers/health/doguChecker_test.go
@@ -46,7 +46,7 @@ func Test_doguChecker_checkDoguHealth(t *testing.T) {
 		ecosystemClientMock := newMockEcosystemInterface(t)
 		ecosystemClientMock.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-		sut := NewDoguChecker(ecosystemClientMock, localFetcher)
+		sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClientMock, localFetcher)
 
 		// when
 		err := sut.CheckByName(testCtx, ldapResource.GetObjectKey())
@@ -66,7 +66,7 @@ func Test_doguChecker_checkDoguHealth(t *testing.T) {
 		ecosystemClient := newMockEcosystemInterface(t)
 		ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-		sut := NewDoguChecker(ecosystemClient, localFetcher)
+		sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 		// when
 		err := sut.CheckByName(testCtx, ldapResource.GetObjectKey())
@@ -88,7 +88,7 @@ func Test_doguChecker_checkDoguHealth(t *testing.T) {
 		ecosystemClient := newMockEcosystemInterface(t)
 		ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-		sut := NewDoguChecker(ecosystemClient, localFetcher)
+		sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 		// when
 		err := sut.CheckByName(testCtx, ldapResource.GetObjectKey())
@@ -152,7 +152,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 		ecosystemClient := newMockEcosystemInterface(t)
 		ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-		sut := NewDoguChecker(ecosystemClient, localFetcher)
+		sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 		// when
 		err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -164,17 +164,17 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 	t.Run("should ignore legacy nginx and registrator dogu dependencies", func(t *testing.T) {
 		/*
 			redmine
-			+-m-> ☑nginx
-			+-m-> ☑registrator
-			+-o-> ☑nginx
-			+-o-> ☑registrator
+			+-m-> ☑️nginx
+			+-m-> ☑️registrator
+			+-o-> ☑️nginx
+			+-o-> ☑️registrator
 		*/
 
 		localFetcher := newMockLocalDoguFetcher(t)
 		ignoreNginxRegistratorDogu := readTestDataDogu(t, ignoreNginxRegistratorBytes)
 		ecosystemClient := newMockEcosystemInterface(t)
 
-		sut := NewDoguChecker(ecosystemClient, localFetcher)
+		sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 		// when
 		err := sut.CheckDependenciesRecursive(testCtx, ignoreNginxRegistratorDogu, testNamespace)
@@ -183,15 +183,99 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 		require.NoError(t, err)
 	})
 
+	t.Run("should ignore legacy cas dependency if authRegistration is enabled", func(t *testing.T) {
+		/*
+			redmine
+			+-m-> ☑️cas
+			+-o-> ☑️cas
+		*/
+
+		localFetcher := newMockLocalDoguFetcher(t)
+		ignoreAuhRegistrationCasDogu := readTestDataDogu(t, ignoreAuhRegistrationCasBytes)
+		ecosystemClient := newMockEcosystemInterface(t)
+
+		sut := NewDoguChecker(&config.OperatorConfig{AuthRegistrationEnabled: true}, ecosystemClient, localFetcher)
+
+		// when
+		err := sut.CheckDependenciesRecursive(testCtx, ignoreAuhRegistrationCasDogu, testNamespace)
+
+		// then
+		require.NoError(t, err)
+	})
+
+	t.Run("should not ignore legacy cas dependency if authRegistration is not enabled", func(t *testing.T) {
+		/*
+			redmine
+			+-m-> ❌️cas
+			+-o-> ❌️cas
+		*/
+
+		localFetcher := newMockLocalDoguFetcher(t)
+		localFetcher.EXPECT().FetchInstalled(testCtx, cescommons.SimpleName("cas")).Return(nil, registryKeyNotFoundTestErr)
+		ignoreAuhRegistrationCasDogu := readTestDataDogu(t, ignoreAuhRegistrationCasBytes)
+		ecosystemClient := newMockEcosystemInterface(t)
+
+		sut := NewDoguChecker(&config.OperatorConfig{AuthRegistrationEnabled: false}, ecosystemClient, localFetcher)
+
+		// when
+		err := sut.CheckDependenciesRecursive(testCtx, ignoreAuhRegistrationCasDogu, testNamespace)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "error fetching local dogu descriptor for dependency \"cas\":")
+	})
+
+	t.Run("should ignore legacy postfix dependency if postfix dependency check is disabled", func(t *testing.T) {
+		/*
+			redmine
+			+-m-> ☑️postfix
+			+-o-> ☑️postfix
+		*/
+
+		localFetcher := newMockLocalDoguFetcher(t)
+		ignorePostfixDependencyDogu := readTestDataDogu(t, ignorePostfixDependencyBytes)
+		ecosystemClient := newMockEcosystemInterface(t)
+
+		sut := NewDoguChecker(&config.OperatorConfig{DisablePostfixDependencyCheck: true}, ecosystemClient, localFetcher)
+
+		// when
+		err := sut.CheckDependenciesRecursive(testCtx, ignorePostfixDependencyDogu, testNamespace)
+
+		// then
+		require.NoError(t, err)
+	})
+
+	t.Run("should not ignore legacy postfix dependency if postfix dependency check is not disabled", func(t *testing.T) {
+		/*
+			redmine
+			+-m-> ❌️postfix
+			+-o-> ❌️postfix
+		*/
+
+		localFetcher := newMockLocalDoguFetcher(t)
+		localFetcher.EXPECT().FetchInstalled(testCtx, cescommons.SimpleName("postfix")).Return(nil, registryKeyNotFoundTestErr)
+		ignorePostfixDependencyDogu := readTestDataDogu(t, ignorePostfixDependencyBytes)
+		ecosystemClient := newMockEcosystemInterface(t)
+
+		sut := NewDoguChecker(&config.OperatorConfig{DisablePostfixDependencyCheck: false}, ecosystemClient, localFetcher)
+
+		// when
+		err := sut.CheckDependenciesRecursive(testCtx, ignorePostfixDependencyDogu, testNamespace)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "error fetching local dogu descriptor for dependency \"postfix\":")
+	})
+
 	t.Run("should ignore client and package dependencies when checking health status of indirect dependencies", func(t *testing.T) {
 		/*
 			testDogu
-			+-m-> ☑ client1 (client)
-			+-m-> ☑ package1 (Package)
-			+-m-> ☑ testDogu2 (Dogu)
-				  +-o-> ☑ client2 (client)
-				  +-o-> ☑ package2 (Package)
-				  +-m-> ☑ testDogu3 (Dogu)
+			+-m-> ☑️ client1 (client)
+			+-m-> ☑️ package1 (Package)
+			+-m-> ☑️ testDogu2 (Dogu)
+				  +-o-> ☑️ client2 (client)
+				  +-o-> ☑️ package2 (Package)
+				  +-m-> ☑️ testDogu3 (Dogu)
 		*/
 
 		testDogu := &core.Dogu{
@@ -226,7 +310,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 		ecosystemClient := newMockEcosystemInterface(t)
 		ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-		sut := NewDoguChecker(ecosystemClient, localFetcher)
+		sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 		// when
 		err := sut.CheckDependenciesRecursive(testCtx, testDogu, testNamespace)
@@ -277,7 +361,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 		ecosystemClient := newMockEcosystemInterface(t)
 		ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-		sut := NewDoguChecker(ecosystemClient, localFetcher)
+		sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 		// when
 		err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -331,7 +415,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -383,7 +467,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -434,7 +518,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -488,7 +572,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -528,7 +612,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -578,7 +662,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -631,7 +715,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -684,7 +768,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -735,7 +819,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -787,7 +871,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -838,7 +922,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)
@@ -854,8 +938,8 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 					redmine
 					+-m-> ☑️postgresql
 					+-m-> ☑️mandatory1
-					+-o-> ☑ optional1
-						  +-m-> ☑ mandatory1
+					+-o-> ☑️ optional1
+						  +-m-> ☑️ mandatory1
 						  +-o-> ~optional2~
 								+-m-> ~mandatory2~
 				*/
@@ -883,7 +967,7 @@ func Test_doguChecker_checkDependencyDogusHealthy(t *testing.T) {
 				ecosystemClient := newMockEcosystemInterface(t)
 				ecosystemClient.EXPECT().Dogus(testNamespace).Return(doguClientMock)
 
-				sut := NewDoguChecker(ecosystemClient, localFetcher)
+				sut := NewDoguChecker(&config.OperatorConfig{}, ecosystemClient, localFetcher)
 
 				// when
 				err := sut.CheckDependenciesRecursive(testCtx, redmineDogu, testNamespace)