+++ controllers/health/testdata_test.go
@@ -39,6 +39,12 @@ var optional2Bytes []byte
 //go:embed testdata/ignore-nginx-registrator.json
 var ignoreNginxRegistratorBytes []byte
 
+//go:embed testdata/ignore-auth-registration-cas.json
+var ignoreAuhRegistrationCasBytes []byte
+
+//go:embed testdata/ignore-postfix-dependency.json
+var ignorePostfixDependencyBytes []byte
+
 func readTestDataLdapCr(t *testing.T) *doguv2.Dogu {
 	t.Helper()